Move get_beneficiary_counts() to authorization
