+++ apps/logging/loggers.py
@@ -4,14 +4,16 @@ from django.core.serializers.json import DjangoJSONEncoder
 import apps.logging.request_logger as logging
 
 from apps.accounts.models import UserProfile
-from apps.authorization.models import get_grant_bene_counts
+from apps.authorization.models import (
+    get_grant_bene_counts,
+    get_beneficiary_counts
+)
 
 from apps.dot_ext.models import (
     Application,
     get_application_counts,
     get_application_require_demographic_scopes_count,
-    get_beneficiary_counts,
-    get_token_bene_counts,
+    get_token_bene_counts
 )
 from apps.fhir.bluebutton.models import get_crosswalk_bene_counts
 from apps.accounts.models import get_developer_counts
